refactor: use a common DeploymnetResourceKind enum
