+++ extensions/kubernetes/vanilla/deployment/src/main/java/io/quarkus/kubernetes/deployment/OpenshiftConfig.java
@@ -1,19 +1,8 @@
 
 package io.quarkus.kubernetes.deployment;
 
-import static io.quarkus.kubernetes.deployment.Constants.BATCH_GROUP;
-import static io.quarkus.kubernetes.deployment.Constants.BATCH_VERSION;
-import static io.quarkus.kubernetes.deployment.Constants.CRONJOB;
-import static io.quarkus.kubernetes.deployment.Constants.DEPLOYMENT;
-import static io.quarkus.kubernetes.deployment.Constants.DEPLOYMENT_CONFIG;
-import static io.quarkus.kubernetes.deployment.Constants.DEPLOYMENT_CONFIG_GROUP;
-import static io.quarkus.kubernetes.deployment.Constants.DEPLOYMENT_CONFIG_VERSION;
-import static io.quarkus.kubernetes.deployment.Constants.DEPLOYMENT_GROUP;
-import static io.quarkus.kubernetes.deployment.Constants.DEPLOYMENT_VERSION;
-import static io.quarkus.kubernetes.deployment.Constants.JOB;
 import static io.quarkus.kubernetes.deployment.Constants.OPENSHIFT;
 import static io.quarkus.kubernetes.deployment.Constants.S2I;
-import static io.quarkus.kubernetes.deployment.Constants.STATEFULSET;
 
 import java.util.Collections;
 import java.util.List;
@@ -37,25 +26,6 @@ public class OpenshiftConfig implements PlatformConfiguration {
     public static enum OpenshiftFlavor {
         v3,
         v4;
-    }
-
-    public static enum DeploymentResourceKind {
-        Deployment(DEPLOYMENT, DEPLOYMENT_GROUP, DEPLOYMENT_VERSION),
-        @Deprecated(since = "OpenShift 4.14")
-        DeploymentConfig(DEPLOYMENT_CONFIG, DEPLOYMENT_CONFIG_GROUP, DEPLOYMENT_CONFIG_VERSION),
-        StatefulSet(STATEFULSET, DEPLOYMENT_GROUP, DEPLOYMENT_VERSION),
-        Job(JOB, BATCH_GROUP, BATCH_VERSION),
-        CronJob(CRONJOB, BATCH_GROUP, BATCH_VERSION);
-
-        public final String kind;
-        public final String apiGroup;
-        public final String apiVersion;
-
-        DeploymentResourceKind(String kind, String apiGroup, String apiVersion) {
-            this.kind = kind;
-            this.apiGroup = apiGroup;
-            this.apiVersion = apiVersion;
-        }
     }
 
     /**
@@ -652,11 +622,10 @@ public class OpenshiftConfig implements PlatformConfiguration {
 
     public DeploymentResourceKind getDeploymentResourceKind(Capabilities capabilities) {
         if (deploymentKind.isPresent()) {
-            return deploymentKind.get();
+            return deploymentKind.filter(k -> k.isAvailalbleOn(OPENSHIFT)).get();
         } else if (capabilities.isPresent(Capability.PICOCLI)) {
             return DeploymentResourceKind.Job;
         }
-
         return (flavor == OpenshiftFlavor.v3) ? DeploymentResourceKind.DeploymentConfig : DeploymentResourceKind.Deployment;
     }
 }
